refactor(web): refactor empty-element into a common component
