+++ web/app/components/app/workflow-log/index.tsx
@@ -1,23 +1,21 @@
 'use client'
-import type { FC, SVGProps } from 'react'
+import type { FC } from 'react'
 import React, { useState } from 'react'
 import useSWR from 'swr'
-import { usePathname } from 'next/navigation'
 import { useDebounce } from 'ahooks'
 import { omit } from 'lodash-es'
 import dayjs from 'dayjs'
 import utc from 'dayjs/plugin/utc'
 import timezone from 'dayjs/plugin/timezone'
-import { Trans, useTranslation } from 'react-i18next'
-import Link from 'next/link'
+import { useTranslation } from 'react-i18next'
 import List from './list'
-import { basePath } from '@/utils/var'
 import Filter, { TIME_PERIOD_MAPPING } from './filter'
+import EmptyElement from '@/app/components/app/log/empty-element'
 import Pagination from '@/app/components/base/pagination'
 import Loading from '@/app/components/base/loading'
 import { fetchWorkflowLogs } from '@/service/log'
 import { APP_PAGE_LIMIT } from '@/config'
-import type { App, AppMode } from '@/types/app'
+import type { App } from '@/types/app'
 import { useAppContext } from '@/context/app-context'
 
 dayjs.extend(utc)
@@ -31,29 +29,6 @@ export type QueryParam = {
   period: string
   status?: string
   keyword?: string
-}
-
-const ThreeDotsIcon = ({ className }: SVGProps<SVGElement>) => {
-  return <svg width="16" height="16" viewBox="0 0 16 16" fill="none" xmlns="http://www.w3.org/2000/svg" className={className ?? ''}>
-    <path d="M5 6.5V5M8.93934 7.56066L10 6.5M10.0103 11.5H11.5103" stroke="#374151" strokeWidth="2" strokeLinecap="round" strokeLinejoin="round" />
-  </svg>
-}
-const EmptyElement: FC<{ appUrl: string }> = ({ appUrl }) => {
-  const { t } = useTranslation()
-  const pathname = usePathname()
-  const pathSegments = pathname.split('/')
-  pathSegments.pop()
-  return <div className='flex h-full items-center justify-center'>
-    <div className='box-border h-fit w-[560px] rounded-2xl bg-background-section-burn px-5 py-4'>
-      <span className='system-md-semibold text-text-secondary'>{t('appLog.table.empty.element.title')}<ThreeDotsIcon className='relative -left-1.5 -top-3 inline' /></span>
-      <div className='system-sm-regular mt-2 text-text-tertiary'>
-        <Trans
-          i18nKey="appLog.table.empty.element.content"
-          components={{ shareLink: <Link href={`${pathSegments.join('/')}/overview`} className='text-util-colors-blue-blue-600' />, testLink: <Link href={appUrl} className='text-util-colors-blue-blue-600' target='_blank' rel='noopener noreferrer' /> }}
-        />
-      </div>
-    </div>
-  </div>
 }
 
 const Logs: FC<ILogsProps> = ({ appDetail }) => {
@@ -78,12 +53,6 @@ const Logs: FC<ILogsProps> = ({ appDetail }) => {
     ...omit(debouncedQueryParams, ['period', 'status']),
   }
 
-  const getWebAppType = (appType: AppMode) => {
-    if (appType !== 'completion' && appType !== 'workflow')
-      return 'chat'
-    return appType
-  }
-
   const { data: workflowLogs, mutate } = useSWR({
     url: `/apps/${appDetail.id}/workflow-app-logs`,
     params: query,
@@ -101,7 +70,7 @@ const Logs: FC<ILogsProps> = ({ appDetail }) => {
           ? <Loading type='app' />
           : total > 0
             ? <List logs={workflowLogs} appDetail={appDetail} onRefresh={mutate} />
-            : <EmptyElement appUrl={`${appDetail.site.app_base_url}${basePath}/${getWebAppType(appDetail.mode)}/${appDetail.site.access_token}`} />
+            : <EmptyElement appDetail={appDetail} />
         }
         {/* Show Pagination only if the total is more than the limit */}
         {(total && total > APP_PAGE_LIMIT)
